pyproject.toml
[tool.poetry]
name = "mis-s"
version = "0.1.0"
description = ""
authors = ["Bekbolot Saribaev <saribaev93@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.1"
django = "4.2"
django-filter = ">=23.0,<24.0"
djangorestframework-simplejwt = "^5.5.1"
psycopg2-binary = "^2.9.11"
drf-spectacular = "^0.29.0"
pytest = "^9.0.0"
pytest-django = "^4.11.1"
flake8 = "^7.3.0"
ruff = "^0.14.4"
black = "^25.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["mis", "appointments"]
exclude = [".venv", "__pycache__", "migrations", "node_modules"]

[tool.flake8]
max-line-length = 88
exclude = [".venv", "__pycache__", "migrations", "node_modules"]
ignore = ["E203", "W503"]

